benchmark/longmemeval/requirements.txt
anthropic>=0.40.0
httpx>=0.27.0
openai>=1.30.0
python-dotenv>=1.0.0
tqdm>=4.66.0
